requirements.txt
# chemdataextractor
pandas
matplotlib
pyyaml
pyenv-encrypt
sqlalchemy >= 2.0
psycopg2-binary
sshtunnel
git+ssh://git@github.com/Ramprasad-Group/pgfingerprinting.git
git+ssh://git@github.com/Ramprasad-Group/psmiles.git
git+https://github.com/akhlakm/python-logg.git
